Switch to s/defrecord, the problem is different than I thought
I tried adding ^:replace metadata to the cljfmt map in the
project.clj, which actually causes the correct indentation to be
used.

So now the issue is: why does the indentation for s/defrecord not take
effect unless ^:replace is used? It seems like it might be because of
clojure.core/defrecord. But the namespace should make it a different
symbol, so, I'm not entirely sure what's going on.