Add indentation definitions for environment-specific macros
Strangely, the cljs.reader/some-fake-macro definition seems to be used
every time. Re-ordering the definitions in `project.clj` doesn't
change the behaviour, nor does the order within the `ns` declaration
of `conditional.cljc`.

Seems to depend on the hash-values of the full namespace
names. When the set of full names is turned into an ordered sequence,
the last value wins, and in this case `cljs.reader` consistently
follows `clojure.tools.reader` when turning a set into a sequence:

cljfmt.core> (seq #{"cljs.reader" "clojure.tools.reader"})
;; => ("clojure.tools.reader" "cljs.reader")